Create README.md
hello Hello is anybody there. Well, I mean, if you are this is gonna be a description of me so yeah. Oh, I've been gaming for about 8 years. And when the mixer app was first launched. I immediately fell in love with it and I am using this app to Edit the mixer broadcasting  To become a better streamer.